Improved bokchoy tests as per code review suggestions.
