+++ common/test/acceptance/tests/lms/test_lms.py
@@ -19,7 +19,6 @@ from ...pages.lms.dashboard import DashboardPage
 from ...pages.lms.problem import ProblemPage
 from ...pages.lms.video.video import VideoPage
 from ...pages.lms.courseware import CoursewarePage
-from ...pages.lms.instructor_dashboard import InstructorDashboardPage
 from ...fixtures.course import CourseFixture, XBlockFixtureDesc, CourseUpdateDesc
 
 
@@ -515,45 +514,3 @@ class ProblemExecutionTest(UniqueCourseTest):
         problem_page.fill_answer("4")
         problem_page.click_check()
         self.assertFalse(problem_page.is_correct())
-
-
-class AutoEnrollmentWithCSVTest(UniqueCourseTest):
-    """
-    Tests of problems.
-    """
-    def setUp(self):
-        super(AutoEnrollmentWithCSVTest, self).setUp()
-        self.course_fixture = CourseFixture(**self.course_info).install()
-
-        # login as an instructor
-        AutoAuthPage(self.browser, course_id=self.course_id, staff=True).visit()
-
-        # go to the membership page on the instructor dashboard
-        self.instructor_dashboard_page = InstructorDashboardPage(self.browser, self.course_id)
-        self.instructor_dashboard_page.visit()
-        self.membership_page = self.instructor_dashboard_page.select_membership()
-
-    def test_browse_and_upload_buttons_are_visible(self):
-        self.assertTrue(self.instructor_dashboard_page.file_attachment_browse_button_is_visible())
-        self.assertTrue(self.instructor_dashboard_page.is_upload_button_visible())
-
-    def test_clicking_file_upload_button_without_file_shows_error(self):
-        self.instructor_dashboard_page.click_upload_file_button()
-        self.assertTrue(self.instructor_dashboard_page.is_error_notification_displayed())
-        self.assertEqual(self.instructor_dashboard_page.first_error_message(), "File is not attached.")
-
-    def test_uploading_correct_csv_file_results_in_success(self):
-        self.instructor_dashboard_page.upload_correct_csv_file()
-        self.assertTrue(self.instructor_dashboard_page.is_success_notification_displayed())
-
-    def test_uploading_csv_file_with_bad_data_results_in_errors_and_warnings(self):
-        self.instructor_dashboard_page.upload_csv_file_with_errors_warnings()
-        self.assertTrue(self.instructor_dashboard_page.is_error_notification_displayed())
-        self.assertTrue(self.instructor_dashboard_page.is_warning_notification_displayed())
-        self.assertEqual(self.instructor_dashboard_page.first_error_message(), "Data in row #2 must have exactly four columns: email, username, full name, and country")
-        self.assertEqual(self.instructor_dashboard_page.first_warning_message(), "ename (d@a.com): (An account with email d@a.com exists but the provided username ename is different. Enrolling anyway with d@a.com.)")
-
-    def test_uploading_non_csv_file_results_in_error(self):
-        self.instructor_dashboard_page.upload_non_csv_file()
-        self.assertTrue(self.instructor_dashboard_page.is_error_notification_displayed())
-        self.assertEqual(self.instructor_dashboard_page.first_error_message(), "Make sure that the file you upload is in CSV format with no extraneous characters or rows.")
